Move ViewDefaultActivityOptions into a constants file
