+++ src/constants/view.ts
@@ -1,0 +1,9 @@
+import type { ActivityOptions } from '../types/timeline';
+
+export const ViewDefaultActivityOptions: ActivityOptions = {
+  activityHeight: 16,
+  composition: 'both',
+  displayMode: 'compact',
+  hierarchyMode: 'flat',
+  labelVisibility: 'auto',
+};
