+++ src/enums/view.ts
@@ -1,14 +1,4 @@
-import type { ActivityOptions } from '../types/timeline';
-
 export enum ViewConstants {
   MIN_ROW_HEIGHT = 24,
   MIN_MARGIN_LEFT = 190,
 }
-
-export const ViewDefaultActivityOptions: ActivityOptions = {
-  activityHeight: 16,
-  composition: 'both',
-  displayMode: 'compact',
-  hierarchyMode: 'flat',
-  labelVisibility: 'auto',
-};
